temporary fix to allow compile with doi issues
Commented out code running QAQC function to allow compile function to execute successfully. Biggest problem is datasets with "israd" entered for doi, I believe. Ultimately this should be handled with additional code but we need a version of this that works today.
Also, I thought I had previously committed these changes but I don't see it on the page. Sorry if I double-committed...