Column additions (4) to master template
After discussions at Irvine, columns were added to the template.

Profile:
pro_chrono_name <- for designating that profile belongs in chronosequence group
pro_soil_age_note <- to allow for designating a profile as a "control" for a chronosequence. Also could allow for noting estimation, expert opinions, etc.
pro_complete <- Whether or not sampling of the entire soil profile was completed, i.e. down to bedrock
pro_NPP <- It was debated whether this could/should be located on the flux tab, but general consensus was that it was more useful/likely to be added on the profile tab