Fix DESCRIPTION file
In commit 83e048087, instead of the version number getting updated, a second version number was added, along with some extraneous lines that look like they came from git. This effectively broke the ability to install the package, either via install_github or locally with install.packages(). This commit deletes the extra lines, and restores functionality to the package