Add files via upload
Data from the disertation of K. Mueller (2010) was originally ingested as Angst_2018 because the former dataset was collected from the same site and used a similar fractionation scheme. However, upon updating the Angst_2018 template to the new fraction tab formatting, it became clear that the 2018 publication does not describe the same data. Rather this 2018 paper reports a similar but different fractionation scheme applied to samples from the same site and Mueller_2010. Thus, it is likely that data from the 2018 paper exists but we will need to reach out to the authors to get the actual dataset since the publication does not include dataset access.