Replaced with Mueller_2010
Data in original Angst_2018 entry do not match the published study and are instead from a related dissertation by Mueller. I've moved the Angst_2018 entry to the In Progress folder. We will need to reach out to the Angst_2018 corresponding author in order to access the appropriate data for that study...