Update QAQC with option to download files
If argument "local" is set to FALSE, the function will download the newest versions of the ISRaD_Master_Template and ISRaD_Template_Info. This was mainly done to add flexibility to Shiny app, so that the QAQC function called there is always using the most up to date input files.